Implementation cupid role
Implemented lovers aspect in their death probabilities

Their amount of werewolf attacks in a single permutation on a lover
is equal to the of the two lovers individually.
This should ensure that whenever a lover is dead in a permutation that
the other lover is also considered dead in that permutation.