performance improvements
the remove() list method in the seer action was the cause for lousy
performance.

To remove this call, the permutations are now represented by a
dictionary indicating wether each permutation is still possible or not.
Fixes #16

The generation of the dictionary is now the bottleneck.
Up to 10 players is now reasonable possible.