Automatic adjustment of table size
The bar length in tables now automatically adjust to the terminal
At the start of the game this length is determined, it is not dynamic.

Under a width of 45 character the minumum of 12 characters in the bar
is used, while for width of 70 and above the width of the bars is fixed
at 36 characters.